legacy_engine/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# OpenAI
openai==1.7.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
newspaper3k==0.2.8

# Database
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
